Make it possible to have two meetups in one month
Our routing setup made it impossible to have two meetups in one month,
because the routing assumed there would only be one event in every
month. This is no longer true, so I changed the individual event routing
to be based on the date (year/month/day), rather than just the month.